* Removing node/test/test-helper.ts added in the 1st commit. * Moving foobarTaskFixture into task-definition-registry.spec.ts.
